+++ backend/src/main/java/dev/salyem/cv/repository/CertificateRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CertificateRepository extends JpaRepository<Certificate, Long> {
+public interface CertificateRepository extends JpaRepository<Certificate, Integer> {
     List<Certificate> findAllByOrderBySortOrderAsc();
 }
