+++ backend/src/main/java/dev/salyem/cv/repository/LanguageRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface LanguageRepository extends JpaRepository<Language, Long> {
+public interface LanguageRepository extends JpaRepository<Language, Integer> {
     List<Language> findAllByOrderBySortOrderAsc();
 }
