+++ backend/src/main/java/dev/salyem/cv/repository/SkillRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SkillRepository extends JpaRepository<Skill, Long> {
+public interface SkillRepository extends JpaRepository<Skill, Integer> {
     List<Skill> findAllByOrderBySortOrderAsc();
     List<Skill> findByCategoryOrderBySortOrderAsc(String category);
 }
